poc7 updated

+++ src/main/java/com/example/criteriaquery/entity/Employee.java
@@ -17,7 +17,7 @@ public class Employee {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long empId;
     private String name;
-    private double salary;
+    private int salary;
 }
